# perf: use HNSW to speed up compute_partitions