--- a/rust/lance-index/src/vector/utils.rs
+++ b/rust/lance-index/src/vector/utils.rs
@@ -5,10 +5,12 @@ use arrow::{
     array::AsArray,
     datatypes::{Float16Type, Float32Type, Float64Type},
 };
-use arrow_array::{Array, ArrowPrimitiveType, BooleanArray, FixedSizeListArray, PrimitiveArray};
+use arrow_array::{Array, ArrayRef, BooleanArray, FixedSizeListArray};
 use arrow_schema::{DataType, Field};
+use lance_arrow::FixedSizeListArrayExt;
 use lance_core::{Error, Result};
 use lance_io::encodings::plain::bytes_to_array;
+use lance_linalg::distance::DistanceType;
 use prost::bytes;
 use snafu::location;
 use std::{ops::Range, sync::Arc};
@@ -20,6 +22,7 @@ use crate::vector::hnsw::builder::HnswBuildParams;
 use crate::vector::hnsw::HNSW;
 use crate::vector::v3::subindex::IvfSubIndex;
 
+#[derive(Debug)]
 pub struct SimpleIndex {
     store: FlatFloatStorage,
     index: HNSW,
@@ -34,11 +37,35 @@ impl SimpleIndex {
         Ok(Self { store, index: hnsw })
     }
 
-    pub(crate) fn search<T: ArrowPrimitiveType>(
-        &self,
-        query: PrimitiveArray<T>,
-    ) -> Result<(u32, f32)> {
-        let query = Arc::new(query);
+    // train HNSW over the centroids to speed up finding the nearest clusters,
+    // only train if all conditions are met:
+    //  - the centroids are float32s or uint8s
+    //  - `num_centroids * dimension >= 1_000_000`
+    //      we benchmarked that it's 2x faster in the case of 1024 centroids and 1024 dimensions,
+    //      so set the threshold to 1_000_000.
+    pub fn may_train_index(
+        centroids: ArrayRef,
+        dimension: usize,
+        distance_type: DistanceType,
+    ) -> Result<Option<Self>> {
+        if centroids.len() < 1_000_000 {
+            // the centroids are stored in a flat array,
+            // the length of the centroids is `num_centroids * dimension`
+            return Ok(None);
+        }
+
+        match centroids.data_type() {
+            DataType::Float32 => {
+                let fsl =
+                    FixedSizeListArray::try_new_from_values(centroids.clone(), dimension as i32)?;
+                let store = FlatFloatStorage::new(fsl, distance_type);
+                Self::try_new(store).map(Some)
+            }
+            _ => Ok(None),
+        }
+    }
+
+    pub(crate) fn search(&self, query: ArrayRef) -> Result<(u32, f32)> {
         let res = self.index.search_basic(query, 1, 10, None, &self.store)?;
         Ok((res[0].id, res[0].dist.0))
     }