… rust/lance-linalg/benches/kmeans.rs
… rust/lance-linalg/src/kmeans.rs