```
compute_membership_1024d_1024k/flat
                        time:   [2.0683 s 2.0825 s 2.0956 s]
compute_membership_1024d_1024k/with_index
                        time:   [896.97 ms 900.17 ms 903.12 ms]

compute_membership_1024d_2048k/flat
                        time:   [7.9075 s 7.9969 s 8.1128 s]
compute_membership_1024d_2048k/with_index
                        time:   [2.0051 s 2.0566 s 2.1118 s]

compute_membership_1024d_4096k/flat
                        time:   [40.691 s 41.660 s 42.597 s]
compute_membership_1024d_4096k/with_index
                        time:   [5.0989 s 5.1553 s 5.2127 s]
```

this is used for training IVF and shuffling,

the e2e indexing time improved from 388.69s to 208.57s (10M rows, 1024d, 2048 partitions)