@BubbleCal 2025-06-27
```
index_params: {'index_type': 'IVF_SQ'}
recall: 99.32%

index_params: {'index_type': 'IVF_HNSW_SQ', 'ef_construction': 10, 'm': 6}
recall: 93.08%

index_params: {'index_type': 'IVF_HNSW_SQ', 'ef_construction': 10, 'm': 12}
recall: 94.74000000000001%

index_params: {'index_type': 'IVF_HNSW_SQ', 'ef_construction': 15, 'm': 6}
recall: 95.81%

index_params: {'index_type': 'IVF_HNSW_SQ', 'ef_construction': 15, 'm': 12}
recall: 98.82%

index_params: {'index_type': 'IVF_HNSW_FLAT', 'ef_construction': 10, 'm': 6}
recall: 93.47999999999999%

index_params: {'index_type': 'IVF_HNSW_FLAT', 'ef_construction': 10, 'm': 12}
recall: 97.45%

index_params: {'index_type': 'IVF_HNSW_FLAT', 'ef_construction': 15, 'm': 6}
recall: 94.65%

index_params: {'index_type': 'IVF_HNSW_FLAT', 'ef_construction': 15, 'm': 12}
recall: 99.49%
```
after testing recall, decided to go with `HNSW_FLAT` with `ef_construction=15` and `m=12`